# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Repo Tokenizer MCP API
  version: 0.1.0
servers:
  - url: http://localhost:4000
paths:
  /health:
    get:
      summary: Health check
      responses:
        '200':
          description: OK
  /index:
    post:
      summary: Trigger repository indexing
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                ref:
                  type: string
      responses:
        '202':
          description: Indexing scheduled
  /files:
    get:
      summary: List indexed files
      parameters:
        - in: query
          name: include
          schema:
            type: string
        - in: query
          name: exclude
          schema:
            type: string
        - in: query
          name: ref
          schema:
            type: string
      responses:
        '200':
          description: List of files
  /file:
    get:
      summary: Get file content
      parameters:
        - in: query
          name: path
          required: true
          schema:
            type: string
        - in: query
          name: ref
          schema:
            type: string
      responses:
        '200':
          description: File payload
  /chunks:
    get:
      summary: List indexed chunks
      parameters:
        - in: query
          name: path
          schema:
            type: string
        - in: query
          name: lang
          schema:
            type: string
        - in: query
          name: maxTokens
          schema:
            type: integer
        - in: query
          name: stream
          schema:
            type: string
        - in: query
          name: ref
          schema:
            type: string
      responses:
        '200':
          description: List of chunks
  /chunks/{id}:
    get:
      summary: Fetch chunk by id
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
        - in: query
          name: ref
          schema:
            type: string
      responses:
        '200':
          description: Chunk payload
  /search:
    get:
      summary: Full-text search within chunks
      parameters:
        - in: query
          name: q
          required: true
          schema:
            type: string
        - in: query
          name: pathGlob
          schema:
            type: string
        - in: query
          name: ref
          schema:
            type: string
      responses:
        '200':
          description: Search results
  /search/symbols:
    get:
      summary: Symbol search
      parameters:
        - in: query
          name: q
          schema:
            type: string
        - in: query
          name: ref
          schema:
            type: string
      responses:
        '200':
          description: Symbol matches
  /export/jsonl:
    get:
      summary: Export current index as JSONL
      responses:
        '200':
          description: JSONL stream
          content:
            application/x-ndjson:
              schema:
                type: string
                format: binary
  /export/sqlite:
    get:
      summary: Export current index as SQLite database
      responses:
        '200':
          description: SQLite binary
          content:
            application/vnd.sqlite3:
              schema:
                type: string
                format: binary